requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
